requirements.txt
arrow
flask
requests
pyyaml
natsort
waitress
jsmin